+++ sdk/src/main/java/ly/count/android/sdk/CountlyWebViewClient.java
@@ -42,4 +42,3 @@ class CountlyWebViewClient extends WebViewClient {
         this.listeners.addAll(listener);
     }
 }
-