+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## XX.XX.XX
-* Added mixed type of immutable lists, arrays to client given segmentations
+* ! Minor breaking change ! Unsupported types for user properties will now be omitted, they won't be converted to strings.
+
+* Added support for mixed type of immutable lists, arrays to client given segmentations and user properties.
+* Added array, list and JSONArray support to the user properties.
+
 * Mitigated issues where:
   * session was ending regardless of manual control after without merge, not anymore.
   * session was not starting even if consent is not required and automatic sessions are enabled after without merge, not anymore.