+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## XX.XX.XX
+* ! Minor breaking change ! Unsupported types for user properties will now be omitted, they won't be converted to strings.
+
+* Added support for mixed type of immutable lists, arrays to client given segmentations and user properties.
+* Added array, list and JSONArray support to the user properties.
+
+* Mitigated issues where:
+  * session was ending regardless of manual control after without merge, not anymore.
+  * session was not starting even if consent is not required and automatic sessions are enabled after without merge, not anymore.
+
 ## 24.7.0
 * ! Minor breaking change ! User properties will now be automatically saved under the following conditions:
   * When an event is recorded